warn fix

+++ text.c
@@ -181,7 +181,7 @@ void draw_number_8x6(uint16_t x, uint16_t y, int32_t number, uint8_t length, uin
 #ifdef SIMULATOR
 	snprintf(s,10, "%i", number);
 #else
-	snprintf(s,10, "%i", number);
+	snprintf(s,10, "%i", (int)number);
 #endif
 
 	int len = strlen(s);
@@ -210,7 +210,7 @@ void draw_number_inv_8x6(uint16_t x, uint16_t y, int32_t number, uint8_t length,
 #ifdef SIMULATOR
 	snprintf(s,10, "%i", number);
 #else
-	snprintf(s,10, "%i", number);
+	snprintf(s,10, "%i", (int)number);
 #endif
 
 	int len = strlen(s);
